Gradient of obj_offspring using the original parameterization.

Gradient of obj_offspring using the original parameterization.

grad_offspring_original(ocounts, osize, ploidy, prob_geno, d, eps, tau)

Arguments

ocounts
The observed counts of the refernce allele for each individual.

osize
The observed number of reads for each individuals.

ploidy
An integer. The ploidy of the species. This is assumed to be the same for each individual.

prob_geno
The allele frequencies of the genotypes. See get_prob_geno.

d
The sequencing bias parameter.

eps
The sequencing error rate

tau
The overdisperion parameter.

Value

A NumericVector of length three with the partial derivatives of d, eps, and tau, in that order.